A selected markups curve gets also selected in the markups module.
When we add marksup curve points, they always concern the last created
curve. If we select a curve in the marksup module, new points are added
to that curve.

So when we select a markups curve in PathReformat, select the same curve
in the markups module.

If None or a VMTK centerline model is selected here, select None in the
markups module. But the last selected curve does not get deselected
there.