Create path bounded ROI.
On creation of a new ROI, it is centered on the selected path, and its bounding
box is strictly that of the path. The edges may still need to be fine tuned.

This is a convenience function. The intent is to quickly prepare an ROI with
minimal editing, based on a path (a markup curve manually created in
particular). It can then be used in other modules like Local Threshold or Crop
Volume.

The typical use case is : axial arterial analysis along a manually created
markup curve, followed by segment creation within the curve bounds in a huge
volume node.

It is not of much value for VMTK centerline models or curves, as they derive
from an existing segment already.